+++ src/Shared/PostToQueue.fs
@@ -4,7 +4,6 @@ open Microsoft.WindowsAzure.Storage
 open FSharp.Control.Tasks.ContextInsensitive
 open Microsoft.WindowsAzure.Storage.Queue
 open FileWriter
-
     module PostToQueue =
 
         type AzureConnection =
@@ -20,21 +19,23 @@ open FileWriter
                             connection.CreateCloudQueueClient()
                         with exn ->
                             let msg =  sprintf "Could not create CloudQueueClient. Message: %s. InnerMessage: %s" exn.Message exn.InnerException.Message
-                            logError exn info msg
+                            // logError exn info msg
+                            Log.logCritical(msg,AzureInfrastucture,Create,QueueTable,exn,info)
+
                             failwith msg
                     let queue =
                         try
                             queueClient.GetQueueReference queueName
                         with exn ->
                             let msg =  sprintf "Could not get Queue Reference. Message: %s. InnerMessage: %s" exn.Message exn.InnerException.Message
-                            logError exn info msg
+                            Log.logCritical(msg,AzureInfrastucture,Create,QueueTable,exn,info)
                             failwith msg
                     let! _q =
                         try
                             queue.CreateAsync()
                         with exn ->
                             let msg =  sprintf "Could not createIfNotExistisAsync Queue. Message: %s. InnerMessage: %s" exn.Message exn.InnerException.Message
-                            logError exn info msg
+                            Log.logCritical(msg,AzureInfrastucture,Create,QueueTable,exn,info)
                             failwith msg
                     return queue
                 }
@@ -43,7 +44,7 @@ open FileWriter
             with
             |exn ->
                 let msg =  sprintf "Could not get Queue. Message: %s. InnerMessage: %s" exn.Message exn.InnerException.Message
-                logError exn info msg
+                Log.logCritical(msg,AzureInfrastucture,Create,QueueTable,exn,info)
                 failwith msg
         let postToQueue (queue:CloudQueue) msg = task {
             let message = CloudQueueMessage(Newtonsoft.Json.JsonConvert.SerializeObject msg)